created console errorHandler
