+++ framework/web/ErrorHandler.php
@@ -62,7 +62,7 @@ class ErrorHandler extends \yii\base\ErrorHandler
 
     /**
      * Renders the exception.
-     * @param \Exception $exception the exception to be handled.
+     * @param \Exception $exception the exception to be rendered.
      */
     protected function renderException($exception)
     {
